ENERGY CONSUMPTION FORECASTING
In this project I will develop a comprehensive analysis of the cnc_machine dataset to explore energy usage patterns and their correlation with CNC machine performance and productivity. Considering that each CNC machine occupies a distinct room of identical dimensions, with each room having the same n7umbers of walls in contact with the outside, my objective is to uncover insights that will help us to identify opportunities for developinmg predictive energy consumption models based on enviroment sensors. In this project I will evaluate the feasability of developing a predictive model tailored to estimate the 10- minute average energy consumption of the factory by utilizing data obtained from sensors within the preceding 1 to 4 hours.